la modularizacion de lo cojone con errores
