+++ app/templates/cuentaAdmins.html
@@ -232,14 +232,14 @@
             <td>
               {% if empleado.5 %}
               <form
-                action="{{ url_for('dar_de_baja_emp', id=empleado[0]) }}"
+                action="{{ url_for('admin.dar_de_baja_emp', id=empleado[0]) }}"
                 method="POST"
               >
                 <input type="submit" value="Dar de baja" />
               </form>
               {% else %}
               <form
-                action="{{ url_for('dar_de_alta_emp', id=empleado[0]) }}"
+                action="{{ url_for('admin.dar_de_alta_emp', id=empleado[0]) }}"
                 method="POST"
               >
                 <input type="submit" value="Dar de alta" />
@@ -282,14 +282,14 @@
             <td>
               {% if cliente.5 %}
               <form
-                action="{{ url_for('dar_de_baja_cli', id=cliente[0]) }}"
+                action="{{ url_for('admin.dar_de_baja_cli', id=cliente[0]) }}"
                 method="POST"
               >
                 <input type="submit" value="Dar de baja" />
               </form>
               {% else %}
               <form
-                action="{{ url_for('dar_de_alta_cli', id=cliente[0]) }}"
+                action="{{ url_for('admin.dar_de_alta_cli', id=cliente[0]) }}"
                 method="POST"
               >
                 <input type="submit" value="Dar de alta" />
@@ -343,7 +343,7 @@
           {% for proveedor in proveedores %}
           <tr data-nombre-prov="{{proveedor.1}}">
             <form
-              action="{{url_for('modify_prov', id=proveedor.0)}}"
+              action="{{url_for('admin.modify_prov', id=proveedor.0)}}"
               method="POST"
               class="form-modify_prov"
               data-id="{{proveedor.0}}"
