+++ app/templates/cuentaEmpleados.html
@@ -306,7 +306,7 @@
             data-nombre-prod="{{ producto.3 }}"
           >
             <form
-              action="{{url_for('modify_prod', id=producto.0)}}"
+              action="{{url_for('employee.modify_prod', id=producto.0)}}"
               method="POST"
               id="form-modify_prod"
               data-id="{{producto.0}}"
@@ -365,14 +365,14 @@
             <td>
               {% if producto.7 %}
               <form
-                action="{{ url_for('dar_de_baja_prod', id=producto[0]) }}"
+                action="{{ url_for('employee.dar_de_baja_prod', id=producto[0]) }}"
                 method="POST"
               >
                 <input type="submit" value="Dar de baja" />
               </form>
               {% else %}
               <form
-                action="{{ url_for('dar_de_alta_prod', id=producto[0]) }}"
+                action="{{ url_for('employee.dar_de_alta_prod', id=producto[0]) }}"
                 method="POST"
               >
                 <input type="submit" value="Dar de alta" />
