+++ app/templates/header.html
@@ -18,7 +18,7 @@
                   <li><a href ="#">Pedidos</a></li>
                   <li>
                     {%if user %}
-                    <a href="#" class="user">
+                    <a href="/cuenta" class="user">
                     <img
                     src="{{url_for('static', filename=user.photo)}}"
                     alt="User"
